0104hsy

+++ src/main/java/com/korea/project2_team4/Service/FollowingMapService.java
@@ -14,16 +14,29 @@ import java.util.List;
 public class FollowingMapService {
     private final FollowingMapRepository followingMapRepository;
 
-    public void setfollowingMap(Profile follower, Profile followee){
-        FollowingMap followingMap = new FollowingMap();
-        followingMap.setFollower(follower);
-        followingMap.setFollowee(followee);
-        this.followingMapRepository.save(followingMap);
+    public void savefollowingMap(Profile follower, Profile followee){
+        if (!followingMapRepository.existsByFollowerAndFollowee(follower, followee)) {
+            FollowingMap followingMap = new FollowingMap();
+            followingMap.setFollower(follower);
+            followingMap.setFollowee(followee);
+            this.followingMapRepository.save(followingMap);
+        } else {
+            // 중복이면 처리할 로직 추가
+            // 예: 이미 존재하는 경우 예외 처리 또는 로깅 등
+            System.out.println("중복된 FollowingMap이 이미 존재합니다.");
+        }
+
     }
 
     public void deletefollowingMap(Profile follower, Profile followee){
-        FollowingMap unfollowingMap = this.followingMapRepository.findByProfiles(follower.getId(), followee.getId());
-        this.followingMapRepository.delete(unfollowingMap);
+        if (followingMapRepository.existsByFollowerAndFollowee(follower, followee)) {
+            FollowingMap unfollowingMap = this.followingMapRepository.findByProfiles(follower.getId(), followee.getId());
+            this.followingMapRepository.delete(unfollowingMap);
+        } else {
+            // 처리할 로직 추가
+            System.out.println("존재하지 않는 FollowingMap입니다.");
+        }
+
     }
 
     public List<Profile> getMyfollowers(Profile profile){
